requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
Pillow>=10.4,<12
python-dotenv>=1.0,<2
openpyxl>=3.1,<4
